feat: Allow running the tool from config file
