+++ tests/fixtures/full_config.yaml
@@ -1,0 +1,24 @@
+backend: clickhouse
+db_host: localhost
+db_port: null
+db_name: xapi_load_test
+db_username: ch_admin
+db_password: test
+run_distribution_queries: false
+drop_tables_first: false
+start_date: 2014-01-01
+end_date: 2023-11-27
+log_dir: logs
+batch_size: 50
+num_batches: 2
+
+num_organizations: 3
+course_sizes:
+  num_huge: 10
+  num_large: 40
+  num_medium: 30
+  num_small: 20
+learner_profiles:
+  num_finishers: 600
+  num_dropoffs: 300
+  num_non_starters: 100
